# students/k3341/Moskalets_Danila/Lr1/teamfinder-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TeamFinder API
  description: Платформа для поиска людей в команду для совместной работы над проектами
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /projects/{project_id}/teams:
    get:
      tags: [teams]
      summary: Получить команды проекта
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список команд с пагинацией
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [teams]
      summary: Создать команду в проекте
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Команда создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teams/{team_id}:
    get:
      tags: [teams]
      summary: Получить команду по ID
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Команда
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [teams]
      summary: Обновить команду (тимлид или владелец проекта)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: Обновлённая команда
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [teams]
      summary: Удалить команду (тимлид или владелец проекта)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Команда удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teams/{team_id}/members:
    get:
      tags: [teams]
      summary: Получить всех участников команды
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [teams]
      summary: Пригласить пользователя в команду
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
      responses:
        '201':
          description: Приглашение отправлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /teams/{team_id}/members/{user_id}:
    patch:
      tags: [teams]
      summary: Обновить статус или роль участника
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRequest'
      responses:
        '200':
          description: Участник обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [teams]
      summary: Удалить участника из команды
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Участник удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── TASKS ───────────────────────────────────────────────
  /projects/{project_id}/tasks:
    get:
      tags: [tasks]
      summary: Получить задачи проекта
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: assignee_id
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/TaskPriority'
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, due_date, priority]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список задач с пагинацией
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [tasks]
      summary: Создать задачу в проекте
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Задача создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Получить задачу по ID
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [tasks]
      summary: Обновить задачу
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Обновлённая задача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [tasks]
      summary: Удалить задачу
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Задача удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── ADMIN ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
      description: ID пользователя
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: integer
      description: ID проекта
    TeamId:
      name: team_id
      in: path
      required: true
      schema:
        type: integer
      description: ID команды
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: integer
      description: ID задачи
  responses:
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, USER]
      description: Роль пользователя
...
    TeamRole:
      type: string
      enum: [LEAD, MEMBER, OBSERVER]
      description: Роль в команде
    MemberStatus:
      type: string
      enum: [INVITED, ACTIVE, LEFT, REMOVED]
      description: Статус участника команды
...
    TaskStatus:
      type: string
      enum: [TODO, IN_PROGRESS, REVIEW, DONE]
      description: Статус задачи
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
      description: Приоритет задачи

    # ── ПАГИНАЦИЯ ──
    Pagination:
      type: object
      required: [total, page, page_size, total_pages]
      properties:
        total:
          type: integer
          description: Общее количество записей
          example: 150
        page:
          type: integer
          description: Текущая страница (1-based)
          minimum: 1
          example: 2
        page_size:
          type: integer
          description: Размер страницы
          minimum: 1
          maximum: 100
          example: 20
        total_pages:
          type: integer
          description: Всего страниц
          example: 8

    # ── AUTH ──
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        role:
          $ref: '#/components/schemas/UserRole'
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@example.com
        avatar_url:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        bio:
          type: string
          nullable: true
          example: Passionate developer
        is_verified:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
...
    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        project_id:
          type: integer
          example: 1
        created_by:
          type: integer
          example: 1
        name:
          type: string
          example: Backend Team
        description:
          type: string
          nullable: true
          example: Responsible for API development
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
    TeamDetailResponse:
      allOf:
        - $ref: '#/components/schemas/TeamResponse'
        - type: object
          properties:
            members:
              type: array
              items:
                $ref: '#/components/schemas/TeamMemberResponse'
    TeamListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TeamResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 256
          example: Frontend Team
        description:
          type: string
          example: UI/UX and frontend development
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          example: Frontend Core Team
        description:
          type: string
          example: Updated description
    TeamMemberResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        role_in_team:
          $ref: '#/components/schemas/TeamRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
        joined_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
    InviteMemberRequest:
      type: object
      required: [user_id, role_in_team]
      properties:
        user_id:
          type: integer
          example: 2
        role_in_team:
          $ref: '#/components/schemas/TeamRole'
    UpdateMemberRequest:
      type: object
      properties:
        role_in_team:
          $ref: '#/components/schemas/TeamRole'
        status:
          $ref: '#/components/schemas/MemberStatus'

    # ── TASKS ──
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        project_id:
          type: integer
          example: 1
        assignee_id:
          type: integer
          nullable: true
          example: 2
        title:
          type: string
          example: Implement authentication
        description:
          type: string
          nullable: true
          example: Add JWT authentication to the API
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date
          nullable: true
          example: 2024-02-15
        created_at:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2024-01-10T00:00:00Z
    TaskListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 256
          example: Setup database
        description:
          type: string
          example: Create PostgreSQL schema
        assignee_id:
          type: integer
          nullable: true
          example: 3
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date
          nullable: true
          example: 2024-01-20
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          example: Update database schema
        description:
          type: string
          nullable: true
          example: Add indexes for performance
        assignee_id:
          type: integer
          nullable: true
          example: 4
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        due_date:
          type: string
          format: date
          nullable: true
          example: 2024-01-25

    # ── ADMIN ──
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: integer
              example: 404
            message:
              type: string
              example: Resource not found
            details:
              type: object
              description: Дополнительные сведения об ошибке
              additionalProperties: true
              nullable: true
...